+++ packages/react-native/src/pages/TripPlanner/TripPlanner.tsx
@@ -1,7 +1,6 @@
 import { useState } from 'react';
 import { View, TouchableOpacity, Dimensions } from 'react-native';
 import { Font, FloatingPlusButton } from 'design-system';
-import { BottomSheetView } from '@gorhom/bottom-sheet';
 import { useNavigation } from '@react-navigation/native';
 import BackGroundGradient from '@/layouts/BackGroundGradient';
 import Header from '@/components/common/Header';
@@ -10,21 +9,10 @@ import useTripPlansQuery, {
   TripPlanResponse,
 } from '@/apis/queries/tripPlan/useTripPlansQuery';
 import TripPlanCard, { CARD_GAP } from '@/components/tripPlan/TripPlanCard';
-import BottomSheet from '@/components/common/BottomSheet';
 import { StackNavigation } from '@/types/navigation';
-import { REGION, REVERSE_REGION_MAPPER } from '@/constants/CITY';
 import withSuspense from '@/components/HOC/withSuspense';
-
-const getDisplayRegion = (selectedPlan: TripPlanResponse) => {
-  const region = REVERSE_REGION_MAPPER[selectedPlan.location];
-  const city = Object.entries(
-    REGION[REVERSE_REGION_MAPPER[selectedPlan.location]],
-  ).find((entry) => {
-    return entry[1] === selectedPlan.city;
-  })?.[0];
-
-  return `${region} ${city}`;
-};
+import EmptyPlan from '@/components/tripPlan/EmptyPlan';
+import TripPlannerBottomSheet from '@/components/tripPlan/TripPlannerBottomSheet';
 
 const { height } = Dimensions.get('window');
 
@@ -63,14 +51,7 @@ export default withSuspense(function TripPlanner() {
             </TouchableOpacity>
           </View>
           {isEmpty ? (
-            <View className="justify-center items-center flex-grow">
-              <Font type="body1" color="white">
-                비어있어요
-              </Font>
-              <Font type="body1" color="white">
-                나만의 여행을 채워보세요
-              </Font>
-            </View>
+            <EmptyPlan />
           ) : (
             <View
               className="mt-5 flex flex-row flex-wrap "
@@ -96,39 +77,7 @@ export default withSuspense(function TripPlanner() {
         right={16}
         onPress={() => navigation.navigate('TripPlanner/Post')}
       />
-      {selectedPlan && (
-        <BottomSheet isShow={Boolean(selectedPlan)} snapPoints={['30%']}>
-          <BottomSheetView
-            style={{
-              flex: 1,
-              justifyContent: 'center',
-            }}
-          >
-            <View className="flex items-center w-full justify-center flex-col">
-              <Font.Bold type="mainTitle" color="black">
-                {getDisplayRegion(selectedPlan)}
-              </Font.Bold>
-              <TouchableOpacity className="py-2" onPress={() => {}}>
-                <Font.Light type="title1" color="black">
-                  여행기간 변경
-                </Font.Light>
-              </TouchableOpacity>
-              <View className="w-[90%] h-[0.5px] bg-[#333333]" />
-              <TouchableOpacity className="py-2">
-                <Font.Light type="title1" color="black">
-                  배너 사진 변경
-                </Font.Light>
-              </TouchableOpacity>
-              <View className="w-[90%] h-[0.5px] bg-[#333333]" />
-              <TouchableOpacity className="py-2" onPress={() => {}}>
-                <Font.Light type="title1" color="black">
-                  삭제
-                </Font.Light>
-              </TouchableOpacity>
-            </View>
-          </BottomSheetView>
-        </BottomSheet>
-      )}
+      <TripPlannerBottomSheet selectedPlan={selectedPlan} />
     </View>
   );
 });
